TreeSuiteBase: move asserting positions
Previously, it was in BasePositionSuite, which also includes parsing.
