+++ tests/shared/src/test/scala/scala/meta/tests/parsers/BasePositionSuite.scala
@@ -1,7 +1,6 @@
 package scala.meta.tests.parsers
 
 import scala.meta.Dialect
-import scala.meta.internal.inputs._
 import scala.meta.parsers.Parse
 
 import munit.Location
@@ -58,56 +57,6 @@ abstract class BasePositionSuite(defaultDialect: Dialect) extends ParseSuite {
     if (expectedTokens.nonEmpty) assertTokenizedAsStructureLines(code.name, expectedTokens)
     val tree = code.name.asInput.parse[T]
       .fold(x => fail("parse failure", x.details), MoreHelpers.requireNonEmptyOrigin(_))
-    val sb = new StringBuilder
-    collect(tree) {
-      // Reduce the expected output by ignoring lines that can be trivially
-      // verified. A line can be trivially verified when you can re-print the
-      // `.syntax` without using tokens. For example, if a Mod.Lazy tree has
-      // the syntax "lazy" then it's trivially verified and excluded from the
-      // output.
-      case `tree` =>
-      case t: Lit.Null if t.text == "null" =>
-      case t: Lit.Unit if t.text == "()" => // This case is needed for Scala.js.
-      case t: Lit if t.value != null && t.text == t.value.toString =>
-      case t: Name if t.text == t.value =>
-      case t @ Importee.Name(Name(value)) if t.text == value =>
-      case t @ Pat.Var(Name(value)) if t.text == value =>
-      case t: Mod if s"Mod.${t.text.capitalize}" == t.productPrefix =>
-      case t: Type.Param if t.text == t.name.value =>
-      case t @ Term.Param(Nil, name, Some(tpe), _) if t.text == s"$name: $tpe" =>
-      case t @ Init(Type.Name(value), anon, Nil) if t.text == value =>
-      case t: Importee.Wildcard if t.text == "_" =>
-      case t: Pat.Wildcard if t.text == "_" =>
-      case t @ Term.ArgClause(arg :: Nil, None) if t.text == arg.text =>
-      case t @ Pat.ArgClause(arg :: Nil) if t.text == arg.text =>
-      case t =>
-        object IterableIndex {
-          def unapply(obj: Any): Option[Int] = obj match {
-            case x: Iterable[_] => x.zipWithIndex.collectFirst { case (`t`, idx) => idx }
-            case _ => None
-          }
-        }
-        val nameOpt =
-          if (showFieldName) t.parent.flatMap(p =>
-            p.productFields.iterator.zip(p.productIterator).collectFirst {
-              case (name, `t`) => name
-              case (name, Some(`t`)) => name
-              case (name, IterableIndex(idx)) => s"$name$idx"
-            }
-          ).orElse(Some("?"))
-          else None
-        nameOpt.foreach(x => sb.append('<').append(x).append('>'))
-        sb.append(t.productPrefix).append(' ')
-        val syntax = t.text
-        if (syntax.isEmpty) {
-          val (leading, trailing) = t.pos.lineContent.splitAt(t.pos.startColumn)
-          sb.append(leading).append("@@").append(trailing)
-        } else sb.append(syntax)
-        nameOpt.foreach(x => sb.append("</").append(x).append('>'))
-        if (showPosition) sb.append(' ').append(t.pos.desc)
-        sb.append('\n')
-    }
-    val obtained = sb.result()
-    assertNoDiff(obtained, expected)
+    assertPositions(tree, expected, showPosition = showPosition, showFieldName = showFieldName)
   }
 }
