+++ tests/shared/src/test/scala/scala/meta/tests/parsers/MoreHelpers.scala
@@ -2,6 +2,7 @@ package scala.meta.tests.parsers
 
 import scala.meta._
 import scala.meta.internal.parsers._
+import scala.meta.tests.TestHelpers
 import scala.meta.tokenizers.TokenizerOptions
 import scala.meta.trees.Origin
 
@@ -9,18 +10,8 @@ import munit._
 
 object MoreHelpers {
 
-  def collect[B](tree: Tree)(pf: PartialFunction[Tree, B]): List[B] = {
-    val builder = List.newBuilder[B]
-    def loop(t: Tree): Unit = {
-      builder ++= pf.lift(t)
-      t.children.foreach(loop)
-    }
-    loop(tree)
-    builder.result()
-  }
-
   def requireNonEmptyOrigin(tree: Tree): tree.type = {
-    val missingOrigin = collect(tree) { case t if t.origin == Origin.None => t }
+    val missingOrigin = TestHelpers.collect(tree) { case t if t.origin == Origin.None => t }
     Assertions.assertEquals(
       missingOrigin.map(_.structure),
       Nil,
